vboot: provide fake nonvolatile storage support
While the actual nonvolatile storage scheme for Storm is being
discussed, it is necessary to provide some plug for vboot to be able
to follow the actual Chrome OS boot sequence.

This fake implementation storage will not survive target reset, but it
is good enough for bringing up the storm target in dev mode.

BUG=chrome-os-partner:27784,chrome-os-partner:29546
TEST=with other patches applied this implementation allows booting of
     Chrome OS on storm

Change-Id: I15bc138afe58b156a2c443b79afce97048388a3c
Reviewed-on: https://chromium-review.googlesource.com/203669
Tested-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Commit-Queue: Vadim Bendebury <vbendeb@chromium.org>