storm: adjust fmap and configuration to match actual system
Currently because of the huge size of the SBL blobs read-only sections
do not fit into 2MB. Rearrange, at lease temporarily, the memory
layout so that all components fit. This, among other things, requires
increasing of the SPI flash cache buffer, and in turn - of the heap
space.

Also, as we do not yet do EFS on this target, there is no need to keep
the 'ramstage' attribute - it confuses cros_bundle_firmware.

BUG=chrome-os-partner:27784
TEST=with this and following patches AP148 boots the Chrome OS kernel.

Change-Id: Idcdc2c06c7ff3a56eb9432ca8415c2d1de43180d
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/203691
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-by: David Hendricks <dhendrix@chromium.org>