devicetree: Change 'coreboot' binding to use a real 'reg' node
On device tree based (= ARM) systems, we currently populate a
'coreboot' compatible node with the two properties 'coreboot-table' and
'coreboot-memory' for the base addresses of these two structures. These
are normal integer properties hardcoded to 32-bit, which is not really
the "right" way to do this.

This patch changes that binding to instead use a normal 'reg' property
(including all the special rules applying to that). The first address
range defined by that property is the coreboot table, the second (which
usually includes the first) is the CBMEM area. Since 'reg' properties
also include the size of an area, the CBMEM information will actually be
useful now (since it's a dynamic, downward-growing structure, parsers
need to start at the end).

This patch only does the minimum required to make things work. There is
a lot of opportunity for refactoring in the device tree code (as
demonstrated by the many existing and new 'TODO' comments), which I will
fix in a later patch that will not be backported into board-specific
branches.

BUG=chrome-os-partner:29311
TEST=Booted a kernel, confirmed that /proc/device-tree looks as
expected. Confirmed that the 'cbmem' utility throws a harmless error
message and doesn't crash violently (it will be fixed later).

Change-Id: I074d345ad63e6d3514b0543549a9024185834639
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/203804
Reviewed-by: David Hendricks <dhendrix@chromium.org>