devicetree: Move coreboot addresses from /firmware to /firmware/coreboot
Olof suggested that using 'ranges' is the lesser evil after all, and
that upstream is more likely to accept it that way. Let's give this
another spin while we (hopeful) still can and pray that it will really
stay final after that.

The kernel driver will still work just as well since it goes by
'compatible', and since the 'ranges' property now allows it to properly
translate the address in the subnode.

BRANCH=nyan
BUG=chrome-os-partner:29311
TEST=Booted, ensured /sys/firmware/log still works (with the other
patch) and /proc/device-tree looks as expected.

Change-Id: Ib2138fede9c654b46cc2f8f0bfb7a49171287a0c
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/203973
Reviewed-by: David Hendricks <dhendrix@chromium.org>