keyboard: Add a timeout when waiting for keys after ESC
This adds a short timeout for when ESC is pressed. If the timoeut
occurs, then assume ESC is the only key we expect.

This implements the suggestion Gabe mentioned in the commit message
in http://crosreview.com/63516.

BUG=chrome-os-partner:29691
BRANCH=nyan
TEST=manually tested by going to dev mode, using ESC to go back after
pressing space or ctrl-d to switch modes, tested that arrow keys work.

Change-Id: I311262104fda8ec66d3b9721f2f37b15a65545fa
Signed-off-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/204393
Reviewed-by: Julius Werner <jwerner@chromium.org>