Provide ability to boot legacy kernels
This is a stripped down copy of the u-boot implementation of the
legacy kernel boot support.

It allows to boot uImage encapsulated arm kernels (using ATAGs as a
means of passing firmware information to the kernel).

It does not yet support any compression, and provides a very limited
number of ATAGs, but is sufficient to boot a legacy kernel on a storm
device.

BUG=chrome-os-partner:27784, chromium:386905, chromium:388011

TEST=with the following patches (enabling this feature and adding a
     CLI command) was able to boot a legacy kernel on storm. See test
     details in the following patch description.

Change-Id: I21887ab78848215de6f3b8576cefb27c535b4445
Reviewed-on: https://chromium-review.googlesource.com/204872
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Reviewed-by: Stefan Reinauer <reinauer@google.com>
Commit-Queue: Vadim Bendebury <vbendeb@chromium.org>
Tested-by: Vadim Bendebury <vbendeb@chromium.org>