Allow to configure legacy kernel support as a second option
depthcharge configuration system allows to enable only a single kernel
type support. Sometimes it is handy to be able to support different
kernel types in the same image.

This patch adds a configuration option for that, making legacy kernel
the only possible second option. A proper fix will involve the ability
to register different kernel loaders at run time, it will be added in
the near future.

BUG=chromium:386905, chromium:388011
TEST=none yet

Change-Id: Ic49b258248afa5add65f8298539675363c1a2e99
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/204873
Reviewed-by: Stefan Reinauer <reinauer@google.com>