Support vendor string based machine id
We want to be able to pass different machine ID to the 3.4 kernel on
ap148 versus proto0 hardware. One way to do it is to key the machine
ID choice by the 'part name' field in the mainboard coreboot table
structure.

In particular, part name 'AP148' uses previously configured value of
4704, and part name 'Storm' uses a new value, 4936. The kernel changes
are to follow.

BUG=chrome-os-partner:30440
TEST=booted kernel on AP148 and proto0, observed expected machine IDs
     reported on the console.

Change-Id: I2edea0fa5681c92e54ca2304906db7f041c5fc0f
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/207458
Reviewed-by: Trevor Bourget <tbourget@codeaurora.org>
Reviewed-by: David Hendricks <dhendrix@chromium.org>