vboot: change VbExEc implementations to support devidx parameter
This will be used in subsequent CLs to support PD software sync.  For
now, only devidx=0 is used, and other values return error.  This makes
static analysis of the change easy.

This must be checked in at the same time as changes to the
vboot_reference project and u-boot implementation.

BUG=chrome-os-partner:30079
BRANCH=none
TEST=emerge-samus depthcharge
CQ-DEPEND=CL:208210,CL:208196

Change-Id: I19c88e6d443841866fced360045b49ea610ce5f5
Signed-off-by: Randall Spangler <rspangler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/208195
Reviewed-by: Bill Richardson <wfrichar@chromium.org>