Refactor out duplicated binary linking code from Makefiles
This patch moves the code that builds the final image back from
vboot/Makefile.inc and netboot/Makefile.inc to a common top level
directory, and replaces the often-duplicated recipes to build images
with a few shared functions. This makes it easier to see which parts
correspond and which are different between multiple images (despite the
unfortunate atrocities that getting variable substitution right in make
functions requires you to commit), and helps in preparation of adding
yet another image type that will combine normal boot and netboot.

BUG=chrome-os-partner:18390
TEST=Built images for Nyan_Blaze, Falco, Daisy and Storm. Confirmed that
all stripped output binaries are byte-for-byte identical.

Change-Id: Ib12bd6a6b9b57712b1b2e186da1ba21abcb0a4d1
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/209270
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>