Add CTRL+N (netboot) to dev images and allow more precise building
Now that we link netboot code into dev images, we are so close to be
able to just offer a "Press CTRL+N for netboot" that it would be silly
not to do it. Kernel folks have been asking for this feature a lot.

Unfortunately, this is still a bit tricky in the details: we cannot just
guard the new key combination with a Kconfig, since those are evaluated
only once (when vboot/callbacks/keyboard.o is built), but we are
building three different images from it. For GDB this worked since we
linked directly to libpayload symbols, but for netboot that is not the
case.

The solution is to move the distinction completely into the linker: we
move all entry points for "debug" functionality into wrapper functions,
and add empty stub implementations with weak symbols for all of them.
This way, production images will recognize CTRL+N and call
dc_dev_netboot(), but that function simply returns without doing
anything... we fall through to the next case and have exactly the same
behavior as if the code hadn't been there at all. In debug builds we
link in the real code for those functions, overriding the weak symbols
and making the key combination work.

We still have the problem that we cannot link to libpayload symbols that
aren't there, which is the case for gdb_enter() and gdb_exit(). Rather
than #ifdef hacking, this patch changes the Makefile to not build any
image by default and require the caller to explicitly build one of the
three image types. With this route we no longer need to build
depthcharge twice from the ebuild... we just build all the code once,
and then link it three times (against different libpayload builds). This
also means we do not need to link both the legacy trampoline image and
the unified EFS image for every board, and can get rid of some hacks
because trampolines do not (seem to?) work on ARMv8.

CQ-DEPEND=CL:209260
BUG=chrome-os-partner:18390
TEST=Boot image.bin, make sure it still boots but CTRL+G and CTRL+N do
nothing. Boot image.dev.bin, make sure it boots normally but can also
netboot (CTRL+N) and enter GDB (CTRL+G/fwgdb). Tested on Nyan_Blaze,
Falco and Daisy to cover ARM+trampoline, ARM+EFS and x86.

Change-Id: I569f417877c23f0ad63ca524c307e31b542afbf7
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/209273
Reviewed-by: Aaron Durbin <adurbin@chromium.org>