ec: Add devidx parameter to cros_ec driver
This supports passthru to the PD chip for the commands needed for
software sync.  It has no effect on other platforms, where devidx=0.
And in truth, it doesn't yet have an effect on samus, because
vboot_reference only ever asks for devidx=0.  It'll start doing
something interesting in a few more CLs.

Tested by verifying that software sync of EC-RW still works.  That is,
adding the devidx parameter should not have regressed current
functionality.

BUG=chrome-os-partner:30079
BRANCH=none
TEST=Flash an old EC-RW image onto samus.
     Flash a new image.bin onto samus.
     Reboot.
     See that software sync takes place properly.

Change-Id: I53626fc6343e52f1777c8c5a39b19f701d083609
Signed-off-by: Randall Spangler <rspangler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/209211
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>