Move common code from our two USB/Ethernet drivers into a separate file
asix.c and smsc95xx.c have a lot of similar code, because they
interact with the same USB and network infrastructure and
do similar things. This change does some refactoring to take
advantage of some of the similarities. More work on the
send(), recv(), and ready() functions might come later.

BUG=None
TEST=Checked that netboot on panther with both asix dongle
and servo still works.

Change-Id: Ief05fc75c208d6f38f32191ab3de57a74b976a47
Signed-off-by: Miriam Gershenson <mgersh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/212265
Reviewed-by: Julius Werner <jwerner@chromium.org>