storm: add ability to force kernel boot from USB for headless devices
Headless devices do not have keyboards and need some other than
pressing '^U' means of triggering kernel boot from a USB stick.

The way to communicate user's desire to boot from USB is pressing the
'recovery' (aka 'physical presence') button *after* the device came
out of reset.

This patch adds another type of flag, to convey the status of this
physical presence signal. When the button is pressed, a '^U' keystroke
is simulated.

BUG=chrome-os-partner:30705, chrome-os-partner:31325

TEST=with the following patch verified USB kernel boot in dev mode
     when the recovery button is pressed after reset.

Change-Id: I68afff0fe228b37e7dd74abc01e1d865f69d7eac
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/212194
Reviewed-by: Randall Spangler <rspangler@chromium.org>