arm64: switch to EL2 in boot_arm_linux_jump()
Currently coreboot is jumping into the payload in EL3 with
caching and the MMU enabled. The previous switch to EL2 was
coopting the page tables set up from coreboot. However, coreboot
is moving the page tables into secure RAM so depthcharge can't
re-use those tables for EL2 as it is inherently unsecure.

Instead just switch to EL2 in the linux kernel boot path. This
sequence can be removed when depthcharge is ran at EL2 from the
get go without any of the MMU enbled.

BUG=chrome-os-partner:31356
BUG=chrome-os-partner:31355
BRANCH=None
TEST=Built and entered the kernel on ryu.

Change-Id: Ibc1ae0a4a016439ed01b5af9c442486f24beea69
Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/213140
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>