spi: add write capability
Upcoming feature development requires ability to write into the SPI
flash. This patch adds the write path in the generic SPI flash driver.

It presumes that the devices are supporting 'legacy' programming mode
when the address size is three bytes, and the well known values are
used for 'write enable' and 'page program' and 'read status register'
commands.

Flash chips do not allow crossing page boundaries during write
transactions. Instead of trying to use the actual page size of a
particular SPI flash device, this patch considers pages to be of 256
bytes in size, which is common to all devices used in coreboot.

BUG=chrome-os-partner:29546
TEST=with the following patch adding 'spi write' CLI command verified
     successful writes including page boundary crossing.

Change-Id: I9640d6e19bfafcce4952f8f383e277f5d7046102
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/213920
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>