Introduce flash storage based NVRAM support
Up to now Chrome OS devices used either CMOS or EC based NVRAM
storage. With the advent of embedded devices none of these are
available, the suggested approach is to store NVRAM blobs in flash
memory along with the rest of Chrome OS firmware.

In this case NVRAM storage uses a fairly large block of flash memory,
which can fit multiple typical NVRAM blobs of 16 bytes. Normally flash
memory allows changing of individual bits from one to zero. Changing
bits from zero to one requires a bulk erase operation, which affects
entire blocks of storage.

This means that the firmware could emulate NVRAM storage using a new
blob in the allocated area each time an NVRAM contents change is
required.

In a typical use case on a running system the last non-erased blob of
CONFIG_NVRAM_BLOCK_SIZE bytes in the dedicated flash block is
considered the current NVRAM contents. If there is a need to change
the NVRAM contents, the next blob worth of bytes is written and then
becomes the last non-erased blob, which is by definition the current
NVRAM contents.

If the entire dedicated block is empty (as when a system boots for the
first time after reprogramming the firmware), the first blob is used
as the NVRAM. It will be considered invalid and overwritten by vboot.

This scheme expects the user space application to manage the allocated
block, erasing it and initializing the lowest blob with the current
nvram contents once the area use gets close to capacity.

If there is no room in the dedicated block for the firmware to store a
new blob - the NVRAM write operation would fail.

The NVRAM area location and size is defined in FMAP.

BUG=chrome-os-partner:29546
TEST=none yet

Change-Id: Id8455f8898569a2fec64b2a01761c0c382495609
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/213960
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>