net: Enable all USB dongle drivers unconditionally
Like all drivers, USB dongles currently need to be activated in the
defconfig. This repeatedly causes issues because drivers are forgotten
when creating or merging a new board.

There is no good reason to ever not activate these options: since they
are USB devices, any dongle can be connected to a port on any board, so
every board with USB support (which is all of them by definition, even
in the latest ChromeOS Core design docs) will want to have them.

BUG=None
TEST=Compiled on Blaze, confirmed that dev images still had ASIX and
SMSC driver code in the binary, and normal images still hadn't.

Change-Id: Ic98b042e37e5adc1e598328481b671b5861f17ba
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/215498
Reviewed-by: Aaron Durbin <adurbin@chromium.org>