storm: configure device tree load address
The storm kernel base address had to be moved to 0x42000000 in order
to leave some space for the NSS firmware. The NSS core is only able to
access low-memory space.

As the kernel can only access the device tree in its mapped area, we also
need to move the device tree to later in the RAM. Here, we're moving it
from 0x41ff0000 to 0x43000000.

BUG=chrome-os-partner:31693
TEST=booted with 3.14 kernel

Change-Id: I1a12454525157ec268276501746af8b9a084d32d
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/215694
Reviewed-by: Bryan Freed <bfreed@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Commit-Queue: Vadim Bendebury <vbendeb@chromium.org>
Tested-by: Vadim Bendebury <vbendeb@chromium.org>