nyan: Expand board ID tables
We just hardcoded a bunch of possible board IDs when we brought up Big
and then never looked at it again. Now we've hit that limit in
production with our frozen RO firmware on Big. -.-

This patch just extends the existing hardcoded table to board ID 10
(0101), which ought to be enough for anybody. Doing it on Blaze and
Kitty as well as a precaution. Note that the hex values it maps to in
depthcharge are *not* simply counting upwards, because we are using this
odd "every tristate digit maps to two binary digits" translation scheme
(instead of ordinary base 3).

We should take this as a warning to not make tristates even more
complicated than they already are and prefer programmatic
board_id-to-revision translators in the future (like on veyron_pinky).

BUG=None
TEST=None (just ship it and pray... or get me a board with flip switches
for the ID).

Change-Id: I9ea7841bed582a49c40f4650a96b1e247cd5965e
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/219526
Reviewed-by: Tom Warren <twarren3959@gmail.com>
Reviewed-by: Katie Roberts-Hoffman <katierh@chromium.org>