fit: Adjust Nyan board IDs to new scheme and clean up compat handling
Since the Nyan board ID representation was changed in coreboot, this
patch needs to change the corresponding depthcharge code to still pick
the right device tree. With the more natural base-3 representation the
values suddenly match up perfectly with kernel -revX numbers, so
generalize the sprintf() approach from Pinky and make it avaiable to
the Nyans. Ryu already did its own translation in coreboot (and seems to
use non-standard kernel DT names), so it stays like it is.

While I'm working on the compat string, also remove the old
KERNEL_FIT_COMPAT Kconfig option and make the "override" mechanism the
new standard. Experience has shown that every shipping board needs to
support multiple DT revs eventually and that the notion of a single
compatible string for the whole overlay isn't very useful... so remove
the useless feature to simplify our logic.

CQ-DEPEND=CL:219901
BUG=None
TEST=See CL:219901

Change-Id: I371f675ca24bbd3f6983fd20ee562ba80b39fddd
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/219902
Reviewed-by: Aaron Durbin <adurbin@chromium.org>