storm: Load the compressed kernel at 0x44000000.
The kernel decompressor is smart enough to detect when decompression
will overwrite its compressed data.  In that case, it moves the
compressed data to just above the end of the uncompressed image.
This happens to work fine for normal kernels (about 3MB compressed),
but for recovery kernels (about 6MB compressed, including initramfs),
we end up overwriting the storm FDT at 0x43000000.

Since the kernel decompressor is released when the actual kernel
starts up, its load address does not affect system memory available
to the kernel.  We might as well choose an address that (1) does not
require movement before decompression, and (2) does not overwrite
the FDT.

BUG=chrome-os-partner:32336
TEST=Confirm we can boot normal and recovery kernels.

Change-Id: I731bbf807722242cad9a1db89d533dbb55355116
Signed-off-by: Bryan Freed <bfreed@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/220112
Reviewed-by: Bill Richardson <wfrichar@chromium.org>