veyron_pinky: Add power button override
Override the sysinfo GPIO definition of the power button with an actual
GPIO, so that it gets sampled at runtime and really works when you press
the button at the developer or recovery screen.

(Note: This is an extremely stupid system. We have all the information
we need in the coreboot tables already, we should just make sysinfo.c
construct actual, run-time sampling GPIOs from it instead of having to
hack overrides like this into every board.)

BUG=None
TEST=Booted rev1 in dev mode. Confirmed it still boots, confirmed it
shuts down when pressing power button.

Change-Id: Iefc539e4d02fbd709c0252a8ada3bc8fafa2b2e2
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/220409
Reviewed-by: David Hendricks <dhendrix@chromium.org>