storm: need larger VPD
New devices will need more room in the VPD to store hardware specific
data. While the exact format is being figured out, this patch
increases the VPD size to accommodate all foreseeable needs.

This increase is at the expense of the GBB, but storm does not need a
large GBB to start with, it is fine to decrease it by 64K.

BUG=chrome-os-partner:32302
TEST=$ dump_fmap -h /build/storm/firmware/image.serial.bin  | grep RO_VPD
    RO_VPD                     003e0000    00400000    00020000

Change-Id: I278fb5b41a6c10be00b7f9f24d43af9fbf3a9b0e
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/221734
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/222571