nand: Streaming interface and GPT on SPI
This patch presents NAND to vboot as a block device allowing reads
and writes to access the GPT (stored on SPI) and allowing a stream to
be opened against a given offset and length. The same interface is
already used for disk as well. The GPT and streaming sections exist
as separate address spaces so that the most natural addresses are
used throughout.

BUG=chromium:403432
TEST=With the following patch, GPT is read correctly from SPI, leading
to a stream being open and verification failing when NAND is erased.
Verified that I didn't break the build by emerge-peppy depthcharge
(a previous version of the patch failed that due to a typo in ahci.c).
TEST=booted from NAND
BRANCH=none
CQ-DEPEND=CL:234640

Signed-off-by: Dan Ehrenberg <dehrenberg@chromium.org>

Change-Id: Ibeee9d6b940120eaa3ec4bf4070000077916e6be
Reviewed-on: https://chromium-review.googlesource.com/222312
Reviewed-by: Julius Werner <jwerner@chromium.org>
Tested-by: Daniel Ehrenberg <dehrenberg@chromium.org>
Commit-Queue: Daniel Ehrenberg <dehrenberg@chromium.org>