storage: Streaming interface and block implementation
This patch adds an implementation in depthcharge for the streaming
interface. A stream can be opened against a disk with a given offset
and length. The change is motivated by NAND, which will have a stream
which skips bad blocks.

BUG=chromium:403432
TEST=Boots from USB on a Storm board when the vboot streaming patch is
included. Verified with print statements that the new streaming read
code is being reached.
BRANCH=none

Signed-off-by: Dan Ehrenberg <dehrenberg@chromium.org>
Change-Id: I6d6826024c3ce2560d6a7a12afc49c28475b234c
Reviewed-on: https://chromium-review.googlesource.com/222648
Reviewed-by: Julius Werner <jwerner@chromium.org>