storm: Boot from NAND
This patch adds a GPT section to Storm's fmap and enables booting from NAND
using the GPT in that location. This is only appropriate for a reworked
Proto 0.2, so before checking this code in, it needs to be refactored into
a separate overlay. Aside from the fmap (which is actually a harmless change),
there's a single config option that needs to be toggled to turn on NAND mode.

TEST=On a Proto 0.2 board with the GPT written to SPI but nothing written to
NAND, this locates the kernel partition, reads all 0xff from NAND, and fails
the verification as it should.
TEST=Booted from NAND on Storm successfully
BUG=chromium:403432
BRANCH=none

Signed-off-by: Dan Ehrenberg <dehrenberg@chromium.org>

Change-Id: I83cc073bf4418177ccdfed630cc1a53a8bd08db7
Reviewed-on: https://chromium-review.googlesource.com/222313
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Daniel Ehrenberg <dehrenberg@chromium.org>
Tested-by: Daniel Ehrenberg <dehrenberg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/237826
Commit-Queue: Vadim Bendebury <vbendeb@chromium.org>
Tested-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>