cosmos: add defconfig and fmap.dts
this adds defconfig and fmap.dts for cosmos. fmap.dts is adjusted
to the size of the spi flash on the board (2MB). the first half is for RO sections
and the second half is for RW sections. read fmap.dts for more details.

BUG=chrome-os-partner:32772
BRANCH=none
TEST=Built chromeos-bootimage
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

Change-Id: If70aea03d38c8d01cefd1b257a67df1e8f6c0477
Reviewed-on: https://chromium-review.googlesource.com/222649
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Commit-Queue: Daisuke Nojiri <dnojiri@chromium.org>
Tested-by: Daisuke Nojiri <dnojiri@chromium.org>