vboot: Add support for new streaming APIs
This implementation is the same as the stub implementations in
vboot_reference, and translates calls to the new streaming APIs to the
old VbExDiskRead() function.  This will allow the vboot_reference
change which uses the new VbExStream*() APIs to commit, which will in
turn allow Dan to debug his real implementation of those new APIs.
Then his implemenation will supersede this stub implementation.

(If we didn't do it in this order, his changes would have to commit
before the vboot code which used them, making testing - and
particularly automated testing in the CQ - more difficult.)

BUG=chromium:403432
BRANCH=none
TEST=The vboot_reference code which calls these APIs hasn't been
     committed yet.  The same code is compiled and tested as part of
     the vboot code unit tests, so it works in another context.  And
     if I add the vboot_reference code, this code still works to boot
     samus.
CQ-DEPEND=CL:221992

Change-Id: Id207e9ae152ed81ca31ed633cf9cd6ac806904da
Signed-off-by: Randall Spangler <rspangler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/222885
Reviewed-by: Daniel Ehrenberg <dehrenberg@chromium.org>