storm: move device tree away from kernel
The placement of the storm device tree is incorrect - it gets
overwritten by the decompressed kernel once the kernel is growing
beyond a certain size.

This patch moves the device tree far away from the kernel to higher
addresses, which is how it is done on other platforms. But the proper
fix should be in the kernel decompressor - it should be intelligent
enough to not overwrite the device tree. It could either stop and
report the problem, or move the device tree somewhere where it would
be out of the way.

Another option would be to place the device tree right above the
loaded kernel blob, not at a fixed address.

BRANCH=storm
BUG=chrome-os-partner:32259
TEST=the previously failing recovery image now boots fine on storm.

Change-Id: I34a646e97961a9f235a86e303a3c7a727fb970bc
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/226346
Reviewed-by: Olof Johansson <olofj@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>