storm: wifi calibration data path does not depend on storage format
The WiFi calibration data read from the VPD and added to the device
tree by deptcharge is put into the device tree node, the name of which
depends on the data format (the '-base64' suffix is added to the node
name). The path should not depend on the storage format, all
calibration data nodes for an interface are expected to be located in
the device tree at the same path.

BRANCH=storm
BUG=chrome-os-partner:32611
TEST=with the new firmware the calibration blobs are placed at the
     correct locations in the device tree:

  localhost soc # vpd -l | cut -c-80
  "wifi_base64_calibration0"="IyAvZXRjL2ZzdGFiOiBzdGF0aWMgZmlsZSBzeXN0ZW0gaW5mb3Jt
  "wifi_base64_calibration1"="IyAvZXRjL2lucHV0cmM6IGluaXRpYWxpemF0aW9uIGZpbGUgZm9y
  localhost soc # find /proc/device-tree  -name '*ath10k-calibration-data*'
  /proc/device-tree/soc/pci@1b700000/pcie@0/ath10k@0,0/qcom,ath10k-calibration-data-base64
  /proc/device-tree/soc/pci@1b500000/pcie@0/ath10k@0,0/qcom,ath10k-calibration-data-base64
  localhost soc #

Change-Id: I2bf61bd20f88e68a91564cc9efcf87fc5749cd48
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/227495
Tested-by: Toshi Kikuchi <toshik@chromium.org>
Reviewed-by: Toshi Kikuchi <toshik@chromium.org>