veyron_jerry: Make board IDs match kernel DT revs 1-to-1
On veryon_pinky, we had a rev0 board without board ID pins and
subsequently called the board with board ID 0 rev1. This required an
offset in the kernel DT compatible string generation. On veyron_jerry,
we don't have any such madness (or rather, we canceled out the madness
of accidentally stuffing the first board ID with 1 by also naming the
first DT rev1), so we can get rid of the offset that was copied over.

BRANCH=none
BUG=None
TEST=None

Change-Id: I76e3c01fbe084f6feddf31d037397c8062ecc174
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/228243
Reviewed-by: David Hendricks <dhendrix@chromium.org>